<feat> RxSwift, Kingfisher, SnapKit 활용 RxSwift input output coordinator Home 화면 구현

+++ Unsplash/Common/Model/Photo.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct Photo: Decodable {
     let id: String
-    let width: Int
-    let height: Int
+    let width: CGFloat
+    let height: CGFloat
     let urls: URLs
     let user: User
 }
